Fixed test

+++ lib/mwlib/tests/MW/Logger/DBTest.php
@@ -52,6 +52,12 @@ class MW_Logger_DBTest extends MW_Unittest_Testcase
 	 */
 	protected function tearDown()
 	{
+		if( TestHelper::getConfig()->get( 'resource/db/adapter', false ) === false ) {
+			return;
+		}
+
+		$this->_dbm = TestHelper::getDBManager();
+
 		$conn = $this->_dbm->acquire();
 
 		$conn->create( 'DROP TABLE "mw_log_test"' )->execute()->finish();
